Process node controller events through a FIFO queue
Prevents nodes from blocking initializations
by filling up the node informer queue with
out of date Update events which still have the
cloud taint on them. The FIFO queue only stores
the latest event for a node, so once a node has
been initialized it will not be initialized again
unless the cloud taint is added back to it.